Refactor Feedback service tests
Lots of clutter in here, and the test names all being the same
makes it difficult to see which tests you have open in an editor.

Streamline test setup (every test has the same setup anyway)
and do some renaming to make it easier to navigate.
