+++ service-api/module/Application/tests/Model/Service/Feedback/FeedbackServiceBuilder.php
@@ -5,7 +5,7 @@ namespace ApplicationTest\Model\Service\Feedback;
 use Application\Model\Service\Feedback\Service;
 use ApplicationTest\Model\Service\AbstractServiceBuilder;
 
-class ServiceBuilder extends AbstractServiceBuilder
+class FeedbackServiceBuilder extends AbstractServiceBuilder
 {
     /**
      * @return Service
